Add session delete by sessionId or key
Adds Session.delete() to delete by provided sessionId or using the
current session key provided with the request.

Adds AuthDB.deleteUserSession() and AuthDB.deleteUserSessionByKey().

Deleting a session using the SessionID will leave a dangling session key
record. The session will still be invalidated, but a cleanup process
will be needed to remove these keys in the future. Since a SessionID
(intentionally) cannot be resolved to it's associated SessionKey, there
is no simple path to do this.